# Improve `loadImage` performance by between 20% and 100%