--- a/src/index.d.ts
+++ b/src/index.d.ts
@@ -145,7 +145,7 @@ declare namespace Tesseract {
     BINARY = 2
   }
   type ImageLike = string | HTMLImageElement | HTMLCanvasElement | HTMLVideoElement
-    | CanvasRenderingContext2D | File | Blob | ImageData | Buffer;
+    | OffscreenCanvas | File | Blob | Buffer;
   interface Block {
     paragraphs: Paragraph[];
     text: string;