> TL;DR: `toBlob()` uses a relatively slow PNG encoder by default. A more uncompressed format like `image/bmp` is better suited for this use case.

## Canvas `.toBlob()` Benchmarks

I tested the performance of `toBlob` with 4 different lossless image formats:

- PNG (the current default, mime type `image/png`)
- RAW (an obscure mime type `image/x-dcraw`)
- TIF (`image/tif`)
- BMP (`image/bmp`)

There's a list of other potential mime types in [this thread](https://stackoverflow.com/a/47612661/16517800), however `image/bmp` is commonly supported and gave me the best results on average (see below).

### On a 1955x3036px Canvas image:

(Using pixel data from the `meditations.jpg` from the `tesseract.js` repo)

```txt
toPngBlob: 442.7470703125 ms
toRawBlob: 360.158935546875 ms
toTifBlob: 362.74609375 ms
toBmpBlob: 354.842041015625 ms
```

Average of **19.9%** speedup for BMP

### On a 100x100px Canvas image:

```txt
toPngBlob: 67.254150390625 ms
toRawBlob: 5.721923828125 ms
toTifBlob: 6.571044921875 ms
toBmpBlob: 1.542724609375 ms
```

Average of **97.7%** speedup for BMP (!!)

### Summary

The benefits of a faster image encoder are most noticeable on a small image size (100x100px or smaller). In addition, Tesseract.js can achieve truly lightning fast speeds on small inputs -- I'm seeing single-digit millisecond calls to `recognize` in some of my testing. It is plenty fast enough for fully realtime use cases when PNG encoding is avoided, since PNG encoding could take 4-5x longer than the actual text recognition at these sizes.

## Other changes

### Add support for `OffscreenCanvas`

This enables Tesseract.js to be started from inside another Web Worker, if desired (for example, after some image pre-processing which already happens off of the main thread). It uses the `OffscreenCanvas.convertToBlob` method which is an exact parallel of `Canvas.toBlob`.

Also, since `HTMLElement` is not defined inside a web worker, I added a check for that before using the `instanceof` method in order to avoid errors when calling `loadImage` from inside a Worker. 

### Update `ImageLike` types

`type ImageLike` currently contains two types for which there is no implementation:

- `CanvasRenderingContext2D`, which is not the intended way to read from `<canvas>` and would have no effect
- `ImageData` which similarly would need to be converted to a blob somehow, and it is easier to start from the `<canvas>` directly

Both options would currently result in runtime errors if they were used. So I've removed these two types, while also adding the `OffscreenCanvas` type as an option.

### (not implemented) Better support for `<video>`

Support for `<video>` sources could be improved, to obtain individual video frames instead of the static `video.poster` which is currently used. This would involve capturing video frames via `canvasContex.drawImage(video, 0, 0)` and then calling `canvas.toBlob()` like any other source. **However**, this would technically be a breaking change if anyone relies on the existing `video.poster` behavior, so I've left that out of this PR for now. I think reading video frames would be a more intuitive behavior than reading the video poster/thumbnail, but for now it would suffice for the user to implement that manually and maintain backwards compatibility.

## Checks

- [X] No breaking changes ✔
- [X] Eslint passing ✔
- [X] `npm run test:all` passes  ✔
- [X] Small change, only a few lines in 1 file + type definitions ✔