@Balearica 2023-04-02
Thank you for this thoughtful contribution and explanation.  Recognizing extremely small images is outside of my personal use case, so I was unaware of the large performance overhead in this specific situation.  My initial thoughts are below. 

1. Can you confirm that your addition of `OffscreenCanvas` support does not cause issues in browsers that do not support this feature? 
   1. Up until very recently `OffscreenCanvas` was only available in Chrome
1. Can you add a unit test for recognizing `OffscreenCanvas`?
   1. We have existing unit tests for other recognition sources, so you should be able to copy/paste and slightly modify an existing test
      1. See https://github.com/naptha/tesseract.js/blob/master/tests/recognize.test.js
1. Can you run the [benchmark code](https://github.com/naptha/tesseract.js/blob/master/examples/browser/benchmark.html) with/without your changes and report the results? 
   1. I want to be sure we have numbers on how this impacts the *entire* runtime of `recognize`, not just the `toBlob` function
   1. Presumably this change won't lead to a big improvement in runtime for the larger test images, but we should confirm that this change does not hurt runtime for certain image types

Additionally, if Tesseract.js switches to using bmp more, I will have to look into this part of the `setImage` function on runtime. 
https://github.com/naptha/tesseract.js/blob/d7d0c2e3c7eb4538c287641506eafba784cbcb61/src/worker-script/utils/setImage.js#L16-L26
At present all bmp images are re-encoded at this step to account for the fact that Leptonica (image processing library used by Tesseract) can not process certain bmp images.  Presumably this adds some overhead.  If this overhead is meaningful, and the bmp images produced by `toBlob` are not subject to these issues, we should think about bypassing this step.