@Balearica 2023-04-03
I looked into this more today and am now confused regarding whether converting from canvas elements to bmp files is supported by browsers at all.  

Running the following in Chrome returns an object where the `type` attribute is `image/png` (the default). 

```
canvas = document.createElement('canvas');
canvas.toBlob((x) => console.log(x))
```

By setting the second argument of `toBlob` to `image/jpeg` the `type` attribute of the object returned changes to `image/jpeg`.

```
canvas.toBlob((x) => console.log(x), "image/jpeg")
```

However, by setting the second argument to `image/bmp` the `type` attribute of the object returned is `image/png`.  This is the same result that occurs when you set the `type` argument to gibberish. 

```
canvas.toBlob((x) => console.log(x), "image/bmp")
```

Additionally, I did not find any indication in [documentation](https://developer.mozilla.org/en-US/docs/Web/API/HTMLCanvasElement/toBlob) for `toBlob` that these image types were supported.  The link you provide merely lists these formats as MIME types, which does not indicate that this particular function is capable of creating images in these formats.