@nathanbabcock 2023-04-04
I just realized a major problem with my benchmark...

I was using the API from `convertToBlob` (the `OffscreenCanvas` variant). That function accepts an **object** like `{ type: 'image/bmp' }` as the second param. But regular canvas `toBlob` takes a plain **string** as the second param. As you noted from your testing, no error or warning gets thrown from an invalid parameter, it just silently falls back to PNG. That means that every test in my benchmark was actually outputting PNG!

But what caused the significant difference in performance then? There must be some kind of caching, because the first call to `toBlob` is always slower than subsequent ones. Because of the way I structured my tests, PNG was the first every single iteration, and took the performance hit. The following three formats all had comparable performance because they were all just cached PNGs. Then I re-rendered the canvas and started again with the next iteration, getting similar results...

Between this and your notes about bitmap weirdness in Tesseract I think might need to rethink the whole approach.

I found another format called PNM in the Leptonica source code.

```c
 /**
  *      The pnm formats are exceedingly simple, because they have
  *      no compression and no colormaps.  They support images that
  *      are 1 bpp; 2, 4, 8 and 16 bpp grayscale; and rgb.
  */
 ```

Source: [`pnmio.c`](https://github.com/DanBloomberg/leptonica/blob/8a5c656cc59032cea72a3a08680616b834f0cfa4/src/pnmio.c)

I assume Tesseract.js uses all the Leptonica stuff under the hood. In theory it would be trivial to take a canvas `ImageData` and just prepend a header to it.