@nathanbabcock 2023-04-16
I've updated the implementation.

---

> 1. Can you confirm that your addition of `OffscreenCanvas` support does not cause issues in browsers that do not support this feature?

Yes. I've guarded each usage of `OffscreenCanvas` by a check like this:

```js
} else if (OffscreenCanvas && image instanceof OffscreenCanvas) {
    // ...
}
```

If `OffscreenCanvas` is not defined in the context (e.g. on older browsers), it will skip over these branches entirely. The same applies in reverse for the usage of `HTMLElement` inside a Web Worker (which has access to `OffscreenCanvas`, but not `HTMLCanvasElement` or `HTMLTag`).

In addition to these fallbacks, `OffscreenCanvas` is supported by all major browsers now: <https://caniuse.com/offscreencanvas>.

---

> 2. Can you add a unit test for recognizing OffscreenCanvas?

Yes, I've added a test that duplicates the regular canvas test exactly, using `OffscreenCanvas`.

---

> Additionally, if Tesseract.js switches to using bmp more, I will have to look into the impact of this part of the setImage function on runtime.

Between this and the questionable behavior of `canvas.toBlob()`, I've rewritten the implementation to convert from canvas `ImageData` directly into a variant of the PBM format. PBM is [supported internally by Tesseract/Leptonica](https://github.com/DanBloomberg/leptonica/blob/master/src/pnmio.c), and is even [covered by the Tesseract.js unit tests already](https://github.com/naptha/tesseract.js/blob/master/tests/constants.js#L18). It can be created instantly from `ImageData` by prepending a simple ASCII-encoded header. The implementation is straightforward:

```js
const imageDataToPBM = (imageData) => {
  const { width, height, data } = imageData;
  const DEPTH = 4; // channels per pixel (RGBA = 4)
  const MAXVAL = 255; // range of each channel (0-255)
  const TUPLTYPE = 'RGB_ALPHA';
  let header = 'P7\n';
  header += `WIDTH ${width}\n`;
  header += `HEIGHT ${height}\n`;
  header += `DEPTH ${DEPTH}\n`;
  header += `MAXVAL ${MAXVAL}\n`;
  header += `TUPLTYPE ${TUPLTYPE}\n`;
  header += 'ENDHDR\n';
  const encoder = new TextEncoder();
  const binaryHeader = encoder.encode(header);
  const binary = new Uint8Array(binaryHeader.length + data.length);
  binary.set(binaryHeader);
  binary.set(data, binaryHeader.length);
  return binary;
};
```