@Balearica 2023-04-17
I tested this today with the [benchmark images](https://github.com/naptha/tesseract.js/tree/master/examples/data), and for the larger images this branch appears to run significantly slower.  For example, when I loaded the largest benchmark image (`meditations.jpg`) to a canvas and then ran recognition this change increased recognition time from 4.4 seconds to 6.2 seconds on Chrome and from 4.7 seconds to 7.8 seconds on Firefox.