@nathanbabcock 2023-04-26
I tried removing the alpha channel. It's a bit faster than before, but still slower than `master`. Here's the comparison (all using `meditations.jpg` on a canvas on Chrome):

## `master`

```txt
loadImage: 452.35400390625 ms
recognize from canvas: 7434.01318359375 ms
```

## `pr`

```txt
loadImage: 143.279052734375 ms
recognize from canvas: 9575.90087890625 ms
```

## `pr` with no alpha channel

```txt
loadImage: 139.8369140625 ms
recognize from canvas: 8361.510986328125 ms
```

I'm pretty confused why `recognize` takes so much longer with this PBM image format. Especially when you can see that `loadImage` runs 4x faster. My assumption was that having immediate access to the raw, uncompressed pixels would be optimal for whatever happens internally inside `tesseract-core`. But this must be mistaken somehow.

Anyone have any insight why raw image data would be slower to recognize than a PNG encoded image?